% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O2PLS_VIP.R
\name{calOrthVIP}
\alias{calOrthVIP}
\title{Calculation of the orthogonal variable influence on projection}
\usage{
calOrthVIP(SSDAO, SSD, loading)
}
\arguments{
\item{SSDAO}{a value of sum of squares (SSDao in step2) for each deflated matrix}

\item{SSD}{the sum of square values}

\item{loading}{the normalized loading matrices}
}
\description{
Calculation of the orthogonal variable influence on projection
}
