% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O2PLS_VIP.R
\name{calPredVIP}
\alias{calPredVIP}
\title{Calculation of the predictive variable influence on projection}
\usage{
calPredVIP(SSXAP, SSYAP, SSD, loading)
}
\arguments{
\item{SSXAP}{the sum of squares values of deflated X matrix for the predictive VIPO2PLS}

\item{SSYAP}{the sum of squares values of deflated Y matrix for the predictive VIPO2PLS}

\item{SSD}{the sum of square values}

\item{loading}{the normalized loading matrices}
}
\description{
Calculation of the predictive variable influence on projection
}
