% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{checkInputdata}
\alias{checkInputdata}
\title{Check if input data satisfies input conditions}
\usage{
checkInputdata(x)
}
\arguments{
\item{x}{A matrix or data frame.}
}
\value{
No return value
}
\description{
This function throws an error if \code{x} is not a numeric matrix or a data
frame with all numeric-alike variables, or if any elements of \code{x} is
\code{NA}.
}
