% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMINP_train.R
\name{get_cvo2mComponent}
\alias{get_cvo2mComponent}
\title{get components number from Cross-validate procedure of O2PLS}
\usage{
get_cvo2mComponent(x)
}
\arguments{
\item{x}{List of class "cvo2m", produced by
\code{\link[OmicsPLS]{crossval_o2m}}.}
}
\value{
A data frame of components number
}
\description{
get components number from Cross-validate procedure of O2PLS
}
