% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldata.R
\docType{data}
\name{MMINP_trained_model}
\alias{MMINP_trained_model}
\title{(Data) An MMINP (O2-PLS) model}
\format{
A list containing an 'o2m' model, results of correlation analysis
between metabolites of training data and its predicted values, and iteration
number of modeling.
}
\description{
This model was built using (\code{\link[MMINP]{MMINP.train}}) with
preprocessed values in dataset \eqn{train_metag} and \eqn{train_metab}.
}
\examples{
data(MMINP_trained_model)
}
\keyword{data}
