% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O2PLS_VIP.R
\name{O2PLSvip}
\alias{O2PLSvip}
\title{Evaluate the importance of variables in O2PLS models}
\usage{
O2PLSvip(x, y, model)
}
\arguments{
\item{x}{Training data of sequence features' relative abundances.
Must have the exact same rows (subjects/samples) as \code{y}.}

\item{y}{Training data of metabolite relative abundances.
Must have the exact same rows (subjects/samples) as \code{x}.}

\item{model}{List of class \code{"mminp"} or \code{"o2m"}, produced by
\code{\link[MMINP]{MMINP.train}} or \code{\link[OmicsPLS]{o2m}}. \code{x}
and \code{y} must be the corresponding training data.}
}
\value{
A list containing
\item{xvip}{For the X-block, the VIP profiles for the predictive part of
the model, the orthogonal part, the total model.}
\item{yvip}{For the Y-block, the VIP profiles for the predictive part of
the model, the orthogonal part, the total model.}
}
\description{
O2PLS-VIP, an approach for variable influence on projection (VIP) in O2PLS
models, is a model-based method for judging the importance of variables. For
both X and Y data blocks, it generates VIP profiles for (i) the predictive
part of the model, (ii) the orthogonal part, and (iii) the total model.
}
\details{
It generates 6 VIPO2PLS profiles in total:
\enumerate{
\item Two VIP profiles for the predictive components, which uncover the X- and
Y-variables that are more important for the model interpretation in relation
to the variation correlated to the Y- and X- data matrices respectively;
\item Two VIP profiles for the orthogonal components for both the X-block and the
Y-block severally, profiles that uncover the X- and Y- variables that are
more relevant in relation to the variation uncorrelated to the Y- and X- data
matrices respectively;
\item Two VIP profiles for the total model (i.e. including the contributions of
both predictive and orthogonal components) for both the X- and the Y- blocks
severally, these VIP profiles point at the X- and Y- variables that are more
significant for the whole model.
}
}
\examples{
#' data(test_metab)
data(test_metag)
a <- MMINP.preprocess(test_metag[, 1:20], normalized = FALSE)
b <- MMINP.preprocess(test_metab[, 1:20], normalized = FALSE)
mminp_model <- MMINP.train(metag = a,
                           metab = b,
                           n = 3:5, nx = 0:3, ny = 0:3,
                           nr_folds = 2, nr_cores = 1)
length(mminp_model$trainres$wellPredicted)
vipres <- O2PLSvip(a, b, mminp_model)
head(vipres$xvip)
head(vipres$yvip)
}
\references{
Galindo-Prieto B, Trygg J, Geladi P. A new approach for variable
influence on projection (VIP) in O2PLS models. Chemometrics and Intelligent
Laboratory Systems 2017; 160: 110–124.
}
