% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMINP_train.R
\name{print.mminp}
\alias{print.mminp}
\title{Print function for MMINP.train}
\usage{
\method{print}{mminp}(x, ...)
}
\arguments{
\item{x}{A model (an object of class "mminp")}

\item{...}{additional parameters}
}
\description{
This function is the print method for \code{MMINP.train}.
}
