% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O2PLS_VIP.R
\name{ssd}
\alias{ssd}
\title{estimation of the sum of squares of deviations}
\usage{
ssd(x)
}
\arguments{
\item{x, }{matrix}
}
\value{
the sum of squares of deviations
}
\description{
estimation of the sum of squares of deviations
}
