% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldata.R
\docType{data}
\name{test_metab}
\alias{test_metab}
\title{(Data) Normalized metabolite abundances for MMINP prediction}
\format{
A list containing two tables of metabolite abundances (proportion
and preprocessed values), with 65 subjects in rows and 130 KEGG compound
IDs in columns.
}
\description{
This datasets were built from NLIBD dataset (Franzosa et al., 2019) by
converting original HMDB IDs into KEGG compound IDs and removing
unassigned and repeated features.
}
\examples{
data(test_metab)
}
\references{
Franzosa EA et al. (2019). Gut microbiome structure and metabolic
activity in inflammatory bowel disease. Nature Microbiology 4(2):293-305.
}
\keyword{data}
