% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldata.R
\docType{data}
\name{train_metab}
\alias{train_metab}
\title{(Data) Normalized metabolite abundances for MMINP training}
\format{
A list containing two tables of metabolite abundances (proportion
and preprocessed values), with 155 subjects in rows and 135 KEGG compound
IDs in columns.
}
\description{
This datasets were built from PRISM dataset (Franzosa et al., 2019) by
converting original HMDB IDs into KEGG compound IDs and removing
unassigned and repeated features.
}
\examples{
data(train_metab)
}
\references{
Franzosa EA et al. (2019). Gut microbiome structure and metabolic
activity in inflammatory bowel disease. Nature Microbiology 4(2):293-305.
}
\keyword{data}
