% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldata.R
\docType{data}
\name{train_metag}
\alias{train_metag}
\title{(Data) Normalized gene family abundances for MMINP training}
\format{
A list containing two tables of gene family abundances (proportion
and preprocessed values), with 155 subjects in rows and 733
KEGG Orthology (KO) IDs in columns.
}
\description{
This datasets were built from PRISM dataset (Franzosa et al., 2019) by
converting original UniRef90 IDs into KEGG Orthology (KO) IDs and removing
unassigned and repeated features.
}
\examples{
data(train_metag)
}
\references{
Franzosa EA et al. (2019). Gut microbiome structure and metabolic
activity in inflammatory bowel disease. Nature Microbiology 4(2):293-305.
}
\keyword{data}
